Unneeded casts from int to float.
The overall expression still returns a float, so the compiler shouldn't cry.

+++ TextFighter/main.cpp
@@ -189,9 +189,9 @@ public:
 void update(Platform p[], Player &player) {
 	world.Step(timeStep, velocityIterations, positionIterations);
 	XInputGetState(0, &state);
-	float x = ((float)state.Gamepad.sThumbLX<1500 && (float)state.Gamepad.sThumbLX>-1500) ? 0 :
+	float x = (state.Gamepad.sThumbLX < 1500 && state.Gamepad.sThumbLX > -1500) ? 0 :
 		((float)state.Gamepad.sThumbLX / 32767);
-	float y = ((float)state.Gamepad.sThumbLY<1500 && (float)state.Gamepad.sThumbLY>-1500) ? 0 :
+	float y = (state.Gamepad.sThumbLY < 1500 && state.Gamepad.sThumbLY > -1500) ? 0 :
 		((float)state.Gamepad.sThumbLY / 32767);
 	for (int i = 0; i < NUMBER_OF_PLATFORMS; i++) {
 		p[i].display();
